// packages/ui-storybook/src/components/button/enums/buttonSizeEnum.ts
import { ButtonTextLarge, ButtonTextMedium, ButtonTextSmall } from '../../typographies';

export enum ButtonWidthSize {
    Large = '48px',
    Medium = '40px',
    Small = '32px'
}

export enum ButtonBorderRadius {
    Large = '24px',
    Medium = '20px',
    Small = '16px'
}

export const ButtonFontFoundation = {
    Large: ButtonTextLarge,
    Medium: ButtonTextMedium,
    Small: ButtonTextSmall,
}

export enum FontButtonSize {
    Large = '16px',
    Medium = '14px',
    Small = '12px'
}

export enum IconButtonSize {
    Large = '24px',
    Medium = '18px',
    Small = '16px'
}

export enum ButtonSizeEnum {
    unset = 'unset',
    Large = 'Large',
    Medium = 'Medium',
    Small = 'Small'
}